feat: update workspace content and removed unused WorkspacePageContent
