--- frontend/src/features/workspace/components/WorkspacePageContent.tsx
@@ -1,83 +1,0 @@
-import { useMemo } from 'react'
-import { Box, Container, Grid } from '@chakra-ui/react'
-
-import { GUIDE_PAYMENTS_ENTRY } from '~constants/links'
-import { ROLLOUT_ANNOUNCEMENT_KEY_PREFIX } from '~constants/localStorage'
-import { useLocalStorage } from '~hooks/useLocalStorage'
-import InlineMessage from '~components/InlineMessage'
-
-// TODO #4279: Remove after React rollout is complete
-import { RolloutAnnouncementModal } from '~features/rollout-announcement/RolloutAnnouncementModal'
-import { useUser } from '~features/user/queries'
-
-import { useWorkspaceContext } from '../WorkspaceContext'
-import { CONTAINER_MAXW } from '../WorkspacePage'
-
-import { EmptyWorkspace } from './EmptyWorkspace'
-import { WorkspaceFormRows } from './WorkspaceFormRow'
-import { WorkspaceHeader } from './WorkspaceHeader'
-
-export interface WorkspacePageContentProps {
-  handleCreateFormModalOpen: () => void
-}
-
-export const WorkspacePageContent = ({
-  handleCreateFormModalOpen,
-}: WorkspacePageContentProps): JSX.Element => {
-  const { user, isLoading: isUserLoading } = useUser()
-
-  const { isLoading, totalFormsCount } = useWorkspaceContext()
-
-  const ROLLOUT_ANNOUNCEMENT_KEY = useMemo(
-    () => ROLLOUT_ANNOUNCEMENT_KEY_PREFIX + user?._id,
-    [user],
-  )
-  const [hasSeenAnnouncement, setHasSeenAnnouncement] =
-    useLocalStorage<boolean>(ROLLOUT_ANNOUNCEMENT_KEY, false)
-
-  const isAnnouncementModalOpen = useMemo(
-    () => !isUserLoading && hasSeenAnnouncement === false,
-    [isUserLoading, hasSeenAnnouncement],
-  )
-
-  const dashboardMessage = `Introducing payments! Citizens can now pay for fees and services directly on your form. [Learn more](${GUIDE_PAYMENTS_ENTRY})`
-
-  return totalFormsCount === 0 ? (
-    <EmptyWorkspace
-      handleOpenCreateFormModal={handleCreateFormModalOpen}
-      isLoading={isLoading}
-    />
-  ) : (
-    <Grid
-      flex={1}
-      overflow="auto"
-      bg="neutral.100"
-      templateColumns="1fr"
-      templateRows="auto auto 1fr"
-      templateAreas="'banner' 'header' 'main'"
-      py="1.5rem"
-    >
-      <Container
-        gridArea="header"
-        maxW={CONTAINER_MAXW}
-        borderBottom="1px solid var(--chakra-colors-neutral-300)"
-        px="2rem"
-        py="1rem"
-      >
-        <InlineMessage useMarkdown mb="2rem" mx="-2rem">
-          {dashboardMessage}
-        </InlineMessage>
-        <WorkspaceHeader
-          handleOpenCreateFormModal={handleCreateFormModalOpen}
-        />
-      </Container>
-      <Box gridArea="main">
-        <RolloutAnnouncementModal
-          onClose={() => setHasSeenAnnouncement(true)}
-          isOpen={isAnnouncementModalOpen}
-        />
-        <WorkspaceFormRows />
-      </Box>
-    </Grid>
-  )
-}
